Sleepy Mode: satisfy cmux-policy-check (docs, file-org, namespace, perf)
- DocC on every new public package symbol (store, config, enums, color hex,
  section body, host actions).
- Allocation-free hex encoding in Color.sleepyHex (drop String(format:),
  PR #5347 class).
- UserDefaults key namespaces are structs, not caseless enums.
- One major type per new file: split the package store file and the five app
  Sleepy files (census, face view, config, controller, power controls) into
  per-type files; wire the new app-target files into the Xcode project.
