+++ Sources/SleepyStar.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+/// One twinkling background star for the Sleepy Mode scene (normalized
+/// position, size, and animation phase/speed).
+struct SleepyStar {
+    let x: Double
+    let y: Double
+    let big: Bool
+    let speed: Double
+    let phase: Double
+}
